Add files via upload
Has some of the new features that Alex asked for: price per mg, vendor name. Working on including the vendors that did not list their prices. There is now a problem with Cayman Chemical - they updated their website so it hides their info with js so we'll have to use Selenium again. Also updated most of the vendor webscraping functions to use built-in functions from BeautifulSoup.